+++ lib/fpm/fry/source/archive.rb
@@ -7,30 +7,62 @@ require 'fpm/fry/source'
 require 'fpm/fry/exec'
 require 'cabin'
 module FPM; module Fry ; module Source
-  class Package
+  # Used to build from an archive.
+  #
+  # @example in a recipe
+  #   source 'http://curl.haxx.se/download/curl-7.36.0.tar.gz',
+  #     checksum: '33015795d5650a2bfdd9a4a28ce4317cef944722a5cfca0d1563db8479840e90'
+  #
+  # It is highly advised to supply a checksum ( althought it's not mandatory ). 
+  # This checksum will be used to test for cache validity and data integrity. The 
+  # checksum algorithm is automatically guessed based on the length of the checksum.
+  #
+  #   - 40 characters = sha1
+  #   - 64 characters = sha256
+  # 
+  # Let's be honest: all other checksum algorithms aren't or shouldn't be in use anyway.
+  class Archive
 
     REGEX = %r!\Ahttps?:!
 
+    # @return [:archive]
     def self.name
       :package
     end
 
     def self.aliases
-      [:http]
+      [:package,:http]
     end
 
+    # Guesses if the given url is an archive.
+    # 
+    # @example not an archive
+    #   FPM::Fry::Source::Archive.guess("bzr://something") # => nil
+    #
+    # @example an archive
+    #   FPM::Fry::Source::Archive.guess("https://some/thing.tar.gz") #=> 6
+    #
+    # @return [nil] when it's not an archive
+    # @return [Numeric] number of characters used
     def self.guess( url )
       Source::guess_regex(REGEX, url)
     end
 
+    # Raised when too many redirects happened.
     class RedirectError < CacheFailed
+    end
+
+    # Raised when the archive type is not known.
+    class UnknownArchiveType < StandardError
+      include WithData
     end
 
     class Cache < Struct.new(:package,:tempdir)
       extend Forwardable
 
-      def_delegators :package, :url, :checksum, :checksum_algorithm, :agent, :logger, :file_map
+      def_delegators :package, :url, :checksum, :checksum_algorithm, :logger, :file_map
 
+      # @return [String] cachekey which is equal to the checksum
       def cachekey
         @observed_checksum || checksum
       end
@@ -93,7 +125,7 @@ module FPM; module Fry ; module Source
           case(resp)
           when Net::HTTPRedirection
             if redirs == 0
-              raise RedirectError, "Too many redirects"
+              raise RedirectError.new("Too many redirects", url: url.to_s, location: resp['location'])
             end
             logger.debug("Following redirect", url: url.to_s , location: resp['location'])
             return fetch_url( resp['location'], redirs - 1, &block)
@@ -167,7 +199,7 @@ module FPM; module Fry ; module Source
         update!
         Exec['unzip', tempfile, '-d', dst ]
       end
-
+    private
       def unpacked_tmpdir
         File.join(tempdir, cachekey)
       end
@@ -198,25 +230,43 @@ module FPM; module Fry ; module Source
       '.bundle' => PlainCache
     }
 
-    attr :file_map, :data, :url, :extension, :checksum, :checksum_algorithm, :agent, :logger
+    attr :file_map, :data, :url, :checksum, :checksum_algorithm, :logger
 
+    # @param [URI] url
+    # @param [Hash] options
+    # @option options [Cabin::Channel] :logger (default cabin channel)
+    # @option options [String] :checksum a checksum of the archive
+    # @raise [UnknownArchiveType] when the archive type is unknown
     def initialize( url, options = {} )
       @url = URI(url)
-      @extension = options.fetch(:extension){
-        CACHE_CLASSES.keys.find{|ext|
-          @url.path.end_with?(ext)
-        }
-      }
+      @cache_class = guess_cache_class(@url)
       @logger = options.fetch(:logger){ Cabin::Channel.get }
       @checksum = options[:checksum]
       @checksum_algorithm = guess_checksum_algorithm(options[:checksum])
       @file_map = options.fetch(:file_map){ {'' => ''} }
     end
 
+    # Creates a cache.
+    #
+    # @param [String] tempdir
+    # @return [TarCache] for plain .tar files
+    # @return [TarGzCache] for .tar.gz files
+    # @return [TarBz2Cache] for .tar.bz2 files
+    # @return [ZipCache] for .zip files
+    # @return [PlainCache] for .bin files
     def build_cache(tempdir)
-      CACHE_CLASSES.fetch(extension).new(self, tempdir)
+      @cache_class.new(self, tempdir)
     end
   private
+
+    def guess_cache_class( url )
+      CACHE_CLASSES.each do |ext,klass|
+        if url.path.end_with?(ext)
+          return klass
+        end
+      end
+      raise UnknownArchiveType.new("Unknown archive type", url: url.to_s, known_extensions: CACHE_CLASSES.keys)
+    end
 
     def guess_checksum_algorithm( checksum )
       case(checksum)