+++ spec/recipe_spec.rb
@@ -242,7 +242,7 @@ RECIPE
     subject{ FPM::Fry::Recipe::Builder.new({}) }
 
     {
-      'http://foo.bar/baz.tar.gz' => FPM::Fry::Source::Package,
+      'http://foo.bar/baz.tar.gz' => FPM::Fry::Source::Archive,
       'http://foo.bar/baz.git'    => FPM::Fry::Source::Git,
       'git@foo.bar:baz/baz.git'   => FPM::Fry::Source::Git,
       'git+ssh://foo.bar/baz'     => FPM::Fry::Source::Git,