+++ spec/source/archive_spec.rb
@@ -1,9 +1,9 @@
-require 'fpm/fry/source/package'
+require 'fpm/fry/source/archive'
 require 'tempfile'
 require 'fileutils'
 require 'webmock/rspec'
 require 'rubygems/package'
-describe FPM::Fry::Source::Package do
+describe FPM::Fry::Source::Archive do
 
   let(:tmpdir){
     Dir.mktmpdir("fpm-fry")
@@ -26,7 +26,7 @@ describe FPM::Fry::Source::Package do
 
     it "fetches a file" do
       stub_request(:get,'http://example/file.tar').to_return(body: "doesn't matter", status: 200)
-      src = FPM::Fry::Source::Package.new("http://example/file.tar")
+      src = FPM::Fry::Source::Archive.new("http://example/file.tar")
       src.build_cache(tmpdir)
       expect( File.read(File.join(tmpdir, 'file.tar')) ).to eq("doesn't matter")
     end
@@ -34,7 +34,7 @@ describe FPM::Fry::Source::Package do
     it "follows redirects" do
       stub_request(:get,'http://example/fileA.tar').to_return(status: 302, headers: {'Location' => 'http://example/fileB.tar'} )
       stub_request(:get,'http://example/fileB.tar').to_return(body: "doesn't matter", status: 200)
-      src = FPM::Fry::Source::Package.new("http://example/fileA.tar")
+      src = FPM::Fry::Source::Archive.new("http://example/fileA.tar")
       src.build_cache(tmpdir)
       expect( File.read(File.join(tmpdir, 'fileA.tar')) ).to eq("doesn't matter")
     end
@@ -42,7 +42,7 @@ describe FPM::Fry::Source::Package do
     it "doesn't follow too many redirects" do
       stub_request(:get,'http://example/fileA.tar').to_return(status: 302, headers: {'Location' => 'http://example/fileB.tar'} )
       stub_request(:get,'http://example/fileB.tar').to_return(status: 302, headers: {'Location' => 'http://example/fileA.tar'} )
-      src = FPM::Fry::Source::Package.new("http://example/fileA.tar")
+      src = FPM::Fry::Source::Archive.new("http://example/fileA.tar")
       expect{
         src.build_cache(tmpdir)
       }.to raise_error( FPM::Fry::Source::CacheFailed, "Too many redirects")
@@ -50,7 +50,7 @@ describe FPM::Fry::Source::Package do
 
     it "reports missing files" do
       stub_request(:get,'http://example/file.tar').to_return(status: 404)
-      src = FPM::Fry::Source::Package.new("http://example/file.tar")
+      src = FPM::Fry::Source::Archive.new("http://example/file.tar")
       expect{
         src.build_cache(tmpdir)
       }.to raise_error(FPM::Fry::Source::CacheFailed, "Unable to fetch file"){|e|
@@ -64,7 +64,7 @@ describe FPM::Fry::Source::Package do
 
     it "reports wrong checksums" do
       stub_request(:get,'http://example/file.tar').to_return(body: "doesn't matter", status: 200)
-      src = FPM::Fry::Source::Package.new("http://example/file.tar", checksum: Digest::SHA256.hexdigest("something else"))
+      src = FPM::Fry::Source::Archive.new("http://example/file.tar", checksum: Digest::SHA256.hexdigest("something else"))
       expect{
         src.build_cache(tmpdir).tar_io
       }.to raise_error(FPM::Fry::Source::CacheFailed, "Checksum failed"){|e|
@@ -77,14 +77,14 @@ describe FPM::Fry::Source::Package do
     end
 
     it "returns checksum as cachekey if present" do
-      src = FPM::Fry::Source::Package.new("http://example/file.tar", checksum: "477c34d98f9e090a4441cf82d2f1f03e64c8eb730e8c1ef39a8595e685d4df65")
+      src = FPM::Fry::Source::Archive.new("http://example/file.tar", checksum: "477c34d98f9e090a4441cf82d2f1f03e64c8eb730e8c1ef39a8595e685d4df65")
       cache = src.build_cache(tmpdir)
       expect( cache.cachekey ).to eq("477c34d98f9e090a4441cf82d2f1f03e64c8eb730e8c1ef39a8595e685d4df65")
     end
 
     it "fetches file for cachekey if no checksum present" do
       stub_request(:get,'http://example/file.tar').to_return(body: "doesn't matter", status: 200)
-      src = FPM::Fry::Source::Package.new("http://example/file.tar")
+      src = FPM::Fry::Source::Archive.new("http://example/file.tar")
       cache = src.build_cache(tmpdir)
       expect( cache.cachekey ).to eq("477c34d98f9e090a4441cf82d2f1f03e64c8eb730e8c1ef39a8595e685d4df65")
     end
@@ -103,7 +103,7 @@ describe FPM::Fry::Source::Package do
 
       it "untars a file" do
         stub_request(:get,'http://example/file.tar').to_return(body: body, status: 200)
-        src = FPM::Fry::Source::Package.new("http://example/file.tar")
+        src = FPM::Fry::Source::Archive.new("http://example/file.tar")
         cache = src.build_cache(tmpdir)
         cache.copy_to(destdir)
         expect( Dir.new(destdir).each.to_a ).to contain_exactly('.','..','foo')
@@ -114,7 +114,7 @@ describe FPM::Fry::Source::Package do
     context '#tar_io' do
       it "untars a file" do
         stub_request(:get,'http://example/file.tar').to_return(body: body, status: 200)
-        src = FPM::Fry::Source::Package.new("http://example/file.tar")
+        src = FPM::Fry::Source::Archive.new("http://example/file.tar")
         cache = src.build_cache(tmpdir)
         expect( cache.tar_io.read ).to eq body
       end
@@ -125,7 +125,7 @@ describe FPM::Fry::Source::Package do
         gz.write(body)
         gz.close
         stub_request(:get,'http://example/file.tar.gz').to_return(body: gzbody.string, status: 200)
-        src = FPM::Fry::Source::Package.new("http://example/file.tar.gz")
+        src = FPM::Fry::Source::Archive.new("http://example/file.tar.gz")
         cache = src.build_cache(tmpdir)
         expect( cache.tar_io.read ).to eq body
       end
@@ -146,7 +146,7 @@ describe FPM::Fry::Source::Package do
     context '#tar_io' do
       it "unzips a zip file" do
         stub_request(:get,'http://example/file.zip').to_return(body: zipfile, status: 200)
-        src = FPM::Fry::Source::Package.new("http://example/file.zip")
+        src = FPM::Fry::Source::Archive.new("http://example/file.zip")
         cache = src.build_cache(tmpdir)
         io = cache.tar_io
         begin
@@ -161,11 +161,11 @@ describe FPM::Fry::Source::Package do
 
       it "doesn't unzip twice" do
         stub_request(:get,'http://example/file.zip').to_return(body: zipfile, status: 200)
-        src = FPM::Fry::Source::Package.new("http://example/file.zip")
+        src = FPM::Fry::Source::Archive.new("http://example/file.zip")
         cache = src.build_cache(tmpdir)
         cache.tar_io.close
 
-        src = FPM::Fry::Source::Package.new("http://example/file.zip")
+        src = FPM::Fry::Source::Archive.new("http://example/file.zip")
         cache = src.build_cache(tmpdir)
         expect(cache).not_to receive(:copy_to)
         cache.tar_io.close
@@ -178,7 +178,7 @@ describe FPM::Fry::Source::Package do
     context '#tar_io' do
       it "tars the file" do
         stub_request(:get,'http://example/dir/plainfile.bin').to_return(body: "bar", status: 200)
-        src = FPM::Fry::Source::Package.new('http://example/dir/plainfile.bin')
+        src = FPM::Fry::Source::Archive.new('http://example/dir/plainfile.bin')
         cache = src.build_cache(tmpdir)
         io = cache.tar_io
         begin
@@ -203,7 +203,7 @@ describe FPM::Fry::Source::Package do
 
       it "copies a file" do
         stub_request(:get,'http://example/dir/plainfile.bin').to_return(body: "bar", status: 200)
-        src = FPM::Fry::Source::Package.new('http://example/dir/plainfile.bin')
+        src = FPM::Fry::Source::Archive.new('http://example/dir/plainfile.bin')
         cache = src.build_cache(tmpdir)
         cache.copy_to(destdir)
         expect( Dir.new(destdir).each.to_a ).to contain_exactly '.','..','plainfile.bin'
@@ -226,7 +226,7 @@ describe FPM::Fry::Source::Package do
     context '#tar_io' do
       it "untars a tar file" do
         stub_request(:get,'http://example/file.tar.bz2').to_return(body: tarfile, status: 200)
-        src = FPM::Fry::Source::Package.new("http://example/file.tar.bz2")
+        src = FPM::Fry::Source::Archive.new("http://example/file.tar.bz2")
         cache = src.build_cache(tmpdir)
         io = cache.tar_io
         begin
@@ -239,5 +239,20 @@ describe FPM::Fry::Source::Package do
       end
 
     end
+  end
+
+  context 'with an unknown extension' do
+
+    it 'raises an error' do
+      expect{
+        FPM::Fry::Source::Archive.new("http://example/file.unknown")
+      }.to raise_error(FPM::Fry::Source::Archive::UnknownArchiveType,"Unknown archive type"){|e|
+        expect(e.data).to include(
+          url: "http://example/file.unknown",
+          known_extensions: include(".tar",".tar.gz")
+        )
+      }
+    end
+
   end
 end